fix(system:destroy): restructure file

+++ commands/system/nuke.sh
@@ -8,6 +8,11 @@ function system:nuke() {
     _logRed "Removing dde sytem"
     system:destroy
 
+    if [ -z ${DATA_DIR} ]; then
+        _logRed "DATA_DIR is not defined"
+        exit 1
+    fi
+
     _logRed "Removing data"
     sudo rm -rf ${DATA_DIR}/*
